Remove prepare and prepare_from_file methods from Filter base class and move them to JaxFilter
