+++ implicit_filter/filter.py
@@ -2,8 +2,6 @@ from abc import ABC, abstractmethod
 from typing import Tuple, List, Union
 
 import numpy as np
-import xarray as xr
-import math
 
 
 class Filter(ABC):
@@ -129,72 +127,3 @@ class Filter(ABC):
     def load_from_file(cls, file: str):
         """Load auxiliary arrays from file"""
         return cls(**dict(np.load(file)))
-
-    @abstractmethod
-    def prepare(self, n2d: int, e2d: int, tri: np.ndarray, xcoord: np.ndarray, ycoord: np.ndarray, meshtype: str = 'r',
-                carthesian: bool = False, cyclic_length: float = 360.0 * math.pi / 180.0, full: bool = False):
-        """
-        Prepare the filter to be used with the given mesh.
-
-        Parameters:
-        -----------
-        n2d : int
-            The total number of nodes in the mesh.
-
-        e2d : int
-            The total number of triangles (elements) in the mesh.
-
-        tri : np.ndarray
-            A 2D NumPy array representing the connectivity of triangles (elements) in the mesh.
-            Each row contains the indices of the three nodes that form a triangle.
-
-        xcoord : np.ndarray
-            A 1D NumPy array containing the x-coordinates of nodes in the mesh.
-
-        ycoord : np.ndarray
-            A 1D NumPy array containing the y-coordinates of nodes in the mesh.
-
-        meshtype : str
-        Mesh type, either 'm' (metric) or 'r' (radial). Default is radial
-
-        carthesian : bool
-            Boolean indicating whether the mesh is in Cartesian coordinates. Default is False
-
-        cyclic_length : float
-            The length of the cyclic boundary if the mesh is cyclic (for 'r' meshtype). Default is 360 * pi / 180
-
-        full : bool, optional
-            A flag indicating whether to use the calculation including metric factors (True) or not (False).
-            Default is False.
-        """
-        pass
-
-    def prepare_from_file(self, file: str, meshtype: str = 'r', carthesian: bool = False,
-                          cyclic_length: float = 360.0 * math.pi / 180.0, metric: bool = False):
-        """
-        Prepare the filter to be used with a mesh provided in the given file path.
-
-        Parameters:
-        -----------
-        file : str
-            Path to the FESOM mesh file.
-
-        meshtype : str
-        Mesh type, either 'm' (metric) or 'r' (radial). Default is radial
-
-        carthesian : bool
-            Boolean indicating whether the mesh is in Cartesian coordinates. Default is False
-
-        cyclic_length : float
-            The length of the cyclic boundary if the mesh is cyclic (for 'r' meshtype). Default is 360 * pi / 180
-
-        metric : bool, optional
-            A flag indicating whether to use the calculation including metric terms (True) or not (False).
-            Default is False.
-        """
-        mesh = xr.open_dataset(file)
-        xcoord = mesh['lon'].values
-        ycoord = mesh['lat'].values
-        tri = mesh['elements'].values.T - 1
-
-        self.prepare(len(xcoord), len(tri[:, 1]), tri, xcoord, ycoord, meshtype, carthesian, cyclic_length, metric)
